% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{alpha_bonf}
\alias{alpha_bonf}
\title{Bonferroni Corrected Alpha}
\usage{
alpha_bonf(alpha, groups)
}
\arguments{
\item{alpha}{significance level}

\item{groups}{number of groups to be post-hos tested}
}
\value{

}
\description{
Bonferroni Corrected Alpha
}
