% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq.R
\name{chisq_explicit}
\alias{chisq_explicit}
\title{Explicit Chi^2 test-value terms}
\usage{
chisq_explicit(observed)
}
\arguments{
\item{observed}{a matrix with observed values}
}
\value{

}
\description{
Each cell holds the value of its contribution to the test statistic. Thus the sum of the test statistic
}
