% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{ci.mean.dev}
\alias{ci.mean.dev}
\title{Mean Confidence Interval Deviation}
\usage{
ci.mean.dev(x, c = 0.95)
}
\arguments{
\item{x}{a data vector}

\item{c}{confidence (between 0..1)}
}
\value{

}
\description{
Mean Confidence Interval Deviation
}
