% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{ci.sd}
\alias{ci.sd}
\title{Standard Deviation Confidence Interval}
\usage{
ci.sd(x, c = 0.95)
}
\arguments{
\item{x}{a data vector}

\item{c}{confidence (between 0..1)}
}
\value{

}
\description{
Standard Deviation Confidence Interval
}
