% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{ci.var}
\alias{ci.var}
\title{Variance Confidence Interval}
\usage{
ci.var(x, c = 0.95)
}
\arguments{
\item{x}{a data vector}

\item{c}{confidence (between 0..1)}
}
\value{

}
\description{
Variance Confidence Interval
}
