% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data_from_count}
\alias{data_from_count}
\title{Data From Count}
\usage{
data_from_count(labels, count)
}
\arguments{
\item{labels}{data labels (can be anytning)}

\item{count}{occurrences of each label}
}
\value{

}
\description{
Creates a data vector with each label appearing the number of times specified in the count vector
}
