% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normplots.R
\name{ecdf_plot}
\alias{ecdf_plot}
\title{ECDF Plotter}
\usage{
ecdf_plot(x)
}
\arguments{
\item{x}{a data vector}
}
\value{

}
\description{
Plot the empirical cumulative distribution vs. the best fitting normal distribution
}
