% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isolve.R
\name{isolve}
\alias{isolve}
\title{Interval Solve}
\usage{
isolve(f, target, from = 0, to = 10000, steps = 10, runs = 20)
}
\arguments{
\item{f}{a function of 1 variable}

\item{target}{the desired output value}

\item{from}{start of search interval (default 0)}

\item{to}{end of search interval (default 10000)}

\item{steps}{number of discrete steps in search interval (default 10)}

\item{runs}{number of recursive calls (default 20)}
}
\value{

}
\description{
Find the x that makes \code{f(x) = target}
}
\examples{
isolve(sqrt, 17.42)
}
