% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{LSD}
\alias{LSD}
\title{Least Significant Difference}
\usage{
LSD(alpha, df, MSE, m)
}
\arguments{
\item{alpha}{bonferroni corrected alpha value}

\item{df}{degrees of freedom in ANOVA}

\item{MSE}{MSE from ANOVA}

\item{m}{number of observations in each treatment group}
}
\value{

}
\description{
Least Significant Difference
}
