% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{mean_ci_convert}
\alias{mean_ci_convert}
\title{Mean Confidence Interval - Converter significance level}
\usage{
mean_ci_convert(ci, df, from = 0.95, to)
}
\arguments{
\item{ci}{confidence interval vector}

\item{df}{degrees of freedom}

\item{from}{initial significance level}

\item{to}{desired significance level}
}
\value{

}
\description{
Mean Confidence Interval - Converter significance level
}
