% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{mean_ci_dev}
\alias{mean_ci_dev}
\title{Mean Confidence Interval Deviation}
\usage{
mean_ci_dev(x, c = 0.95)
}
\arguments{
\item{x}{a data vector}

\item{c}{confidence (between 0..1)}
}
\value{

}
\description{
Mean Confidence Interval Deviation
}
