% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normplots.R
\name{qq_plot}
\alias{qq_plot}
\title{QQ Plotter}
\usage{
qq_plot(x, padding = 0.2)
}
\arguments{
\item{x}{a data vector}

\item{padding}{amount of extra vertical range (default = 0.2)}
}
\value{

}
\description{
Makes a QQ-plot that can easily be adjusted to look nice
}
