% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd.R
\name{sd_ci}
\alias{sd_ci}
\title{Standard Deviation Confidence Interval}
\usage{
sd_ci(x, c = 0.95)
}
\arguments{
\item{x}{a data vector}

\item{c}{confidence (between 0..1)}
}
\value{

}
\description{
Standard Deviation Confidence Interval
}
