% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd.R
\name{sd_ci_nv}
\alias{sd_ci_nv}
\title{Standard Deviation Confidence Interval - No Vector}
\usage{
sd_ci_nv(s, n, c = 0.95)
}
\arguments{
\item{s}{standard deviation}

\item{n}{number of observations}

\item{c}{confidence (between 0..1)}
}
\value{

}
\description{
Standard Deviation Confidence Interval - No Vector
}
