% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd.R
\name{sd_scale}
\alias{sd_scale}
\title{Standard Deviation Scaling}
\usage{
sd_scale(sd, k)
}
\arguments{
\item{sd}{a given standard deviation}

\item{k}{scaling constant}
}
\value{

}
\description{
Standard Deviation Scaling
}
