% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t.R
\name{t_obs_2s_welch}
\alias{t_obs_2s_welch}
\title{Two Sample t-value (Welch Method)}
\usage{
t_obs_2s_welch(x1, x2, s1, s2, n1, n2, delta0 = 0)
}
\arguments{
\item{x1}{sample mean group 1}

\item{x2}{sample mean group 2}

\item{s1}{sample sd group 1}

\item{s2}{sample sd group 2}

\item{n1}{number of samples in group 1}

\item{n2}{number of samples in group 2}

\item{delta0}{offset}
}
\value{

}
\description{
Two Sample t-value (Welch Method)
}
