% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.R
\name{var_ci}
\alias{var_ci}
\title{Variance Confidence Interval}
\usage{
var_ci(x, c = 0.95)
}
\arguments{
\item{x}{a data vector}

\item{c}{confidence (between 0..1)}
}
\value{

}
\description{
Variance Confidence Interval
}
