% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.R
\name{var_ci_nv}
\alias{var_ci_nv}
\title{Variance Confidence Interval - No Vector}
\usage{
var_ci_nv(s, n, c = 0.95)
}
\arguments{
\item{s}{standard deviation}

\item{n}{number of observations}

\item{c}{confidence (between 0..1)}
}
\value{

}
\description{
Variance Confidence Interval - No Vector
}
