% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.R
\name{var_percent}
\alias{var_percent}
\title{Variance From a Percentage Distribution}
\usage{
var_percent(values, distribution)
}
\arguments{
\item{values}{a vector of values}

\item{distribution}{a vector representing the relative distribution of the values}
}
\value{

}
\description{
computes the variance given only the relative distribution between a set of values
}
