% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z.R
\name{z_obs_1}
\alias{z_obs_1}
\title{Z-obs for 1 proportion (Method 7.11)}
\usage{
z_obs_1(x, n, p0)
}
\arguments{
\item{x}{proportion count}

\item{n}{total count}

\item{p0}{proportion hypothesis}
}
\value{

}
\description{
Z-obs for 1 proportion (Method 7.11)
}
