% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{mean_ci_nv}
\alias{mean_ci_nv}
\title{Mean Confidence Interval - No Vector}
\usage{
mean_ci_nv(mu, s, n, c = 0.95)
}
\arguments{
\item{mu}{mean}

\item{s}{standard deviation}

\item{n}{degrees of freedom}

\item{c}{confidence (between 0..1)}
}
\value{

}
\description{
Mean Confidence Interval - No Vector
}
\examples{
mean_ci_nv(mu = -0.1181818, s = 0.05884899, n = 22)
}
